requirements.txt
streamlit 
OpenAI
langchain
python-dotenv
langchain_nvidia_ai_endpoints